+++ src/core/functions/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(scalar)
 
 set(EXTENSION_SOURCES
-    ${EXTENSION_SOURCES}
+    ${EXTENSION_SOURCES} ${CMAKE_CURRENT_SOURCE_DIR}/prompt_builder.cpp
     PARENT_SCOPE)
