+++ src/core/functions/scalar/llm_complete_json.cpp
@@ -1,6 +1,6 @@
 #include <functional>
-#include <inja/inja.hpp>
 #include <iostream>
+#include <flockmtl/core/functions/prompt_builder.hpp>
 #include <flockmtl/common.hpp>
 #include <flockmtl/core/functions/scalar.hpp>
 #include <flockmtl/core/model_manager/model_manager.hpp>
@@ -16,93 +16,6 @@
 
 namespace flockmtl {
 namespace core {
-
-template <typename T>
-std::string ToString(const T &value) {
-    std::ostringstream oss;
-    oss << value;
-    return oss.str();
-}
-
-nlohmann::json GetMaxLengthValues(const std::vector<nlohmann::json> &params) {
-    nlohmann::json attr_to_max_token_length;
-
-    for (const auto &json_obj : params) {
-        for (const auto &item : json_obj.items()) {
-            auto key = item.key();
-            auto value_str = ToString(item.value());
-            int length = value_str.length();
-
-            if (attr_to_max_token_length.contains(key)) {
-                auto current_max_value_str = ToString(attr_to_max_token_length[key]);
-                if (current_max_value_str.length() < length) {
-                    attr_to_max_token_length[key] = item.value();
-                }
-            } else {
-                attr_to_max_token_length[key] = item.value();
-            }
-        }
-    }
-
-    return attr_to_max_token_length;
-}
-
-std::string combine_values(const nlohmann::json &json_obj) {
-    std::string combined;
-    for (const auto &item : json_obj.items()) {
-        combined += ToString(item.value()) + " ";
-    }
-
-    if (!combined.empty()) {
-        combined.pop_back();
-    }
-    return combined;
-}
-
-inline std::vector<std::string> ConstructPrompts(std::vector<nlohmann::json> &unique_rows, Connection &con,
-                                                 std::string prompt_name, int model_max_tokens = 4096) {
-    inja::Environment env;
-
-    auto query_result = con.Query(
-        "SELECT prompt FROM flockmtl_config.FLOCKMTL_PROMPT_INTERNAL_TABLE WHERE prompt_name = '" + prompt_name + "'");
-
-    if (query_result->RowCount() == 0) {
-        throw std::runtime_error("Prompt not found");
-    }
-
-    auto template_str = query_result->GetValue(0, 0).ToString();
-    auto row_tokens = Tiktoken::GetNumTokens(template_str);
-    auto max_length_values = GetMaxLengthValues(unique_rows);
-    auto combined_values = combine_values(max_length_values);
-    row_tokens += Tiktoken::GetNumTokens(combined_values);
-
-    std::vector<std::string> prompts;
-
-    if (row_tokens > model_max_tokens) {
-        throw std::runtime_error("The total number of tokens in the prompt exceeds the model's maximum token limit");
-    } else {
-        auto template_tokens = Tiktoken::GetNumTokens(llm_complete_json_prompt_template);
-        auto max_tokens_for_rows = model_max_tokens - template_tokens;
-        auto max_chunk_size = max_tokens_for_rows / row_tokens;
-        auto chunk_size = max_chunk_size > static_cast<int>(unique_rows.size()) ? static_cast<int>(unique_rows.size())
-                                                                                : max_chunk_size;
-        auto num_chunks = static_cast<int>(std::ceil(static_cast<double>(unique_rows.size()) / chunk_size));
-
-        for (int i = 0; i < num_chunks; ++i) {
-            nlohmann::json data;
-            data["prompts"] = template_str;
-
-            for (int j = 0; j < chunk_size; ++j) {
-                data["rows"].push_back(unique_rows[i + j]);
-            }
-
-            auto prompt = env.render(llm_complete_json_prompt_template, data);
-            prompts.push_back(prompt);
-        }
-    }
-
-    return prompts;
-}
 
 static void LlmCompleteJsonScalarFunction(DataChunk &args, ExpressionState &state, Vector &result) {
     Connection con(*state.GetContext().db);
@@ -137,7 +50,8 @@ static void LlmCompleteJsonScalarFunction(DataChunk &args, ExpressionState &stat
     } else {
         auto tuples = CoreScalarParsers::Struct2Json(args.data[2], args.size());
 
-        auto prompts = ConstructPrompts(tuples, con, args.data[0].GetValue(0).ToString(), model_max_tokens);
+        auto prompts = ConstructPrompts(tuples, con, args.data[0].GetValue(0).ToString(),
+                                        llm_complete_json_prompt_template, model_max_tokens);
 
         nlohmann::json settings;
         if (args.ColumnCount() == 4) {
