+++ src/core/functions/scalar/llm_filter.cpp
@@ -1,6 +1,6 @@
 #include <functional>
-#include <inja/inja.hpp>
 #include <iostream>
+#include <flockmtl/core/functions/prompt_builder.hpp>
 #include <flockmtl/common.hpp>
 #include <flockmtl/core/functions/scalar.hpp>
 #include <flockmtl/core/model_manager/model_manager.hpp>
@@ -16,94 +16,7 @@
 namespace flockmtl {
 namespace core {
 
-template <typename T>
-std::string ToString2(const T &value) {
-    std::ostringstream oss;
-    oss << value;
-    return oss.str();
-}
-
-nlohmann::json GetMaxLengthValues2(const std::vector<nlohmann::json> &params) {
-    nlohmann::json attr_to_max_token_length;
-
-    for (const auto &json_obj : params) {
-        for (const auto &item : json_obj.items()) {
-            auto key = item.key();
-            auto value_str = ToString2(item.value());
-            int length = value_str.length();
-
-            if (attr_to_max_token_length.contains(key)) {
-                auto current_max_value_str = ToString2(attr_to_max_token_length[key]);
-                if (current_max_value_str.length() < length) {
-                    attr_to_max_token_length[key] = item.value();
-                }
-            } else {
-                attr_to_max_token_length[key] = item.value();
-            }
-        }
-    }
-
-    return attr_to_max_token_length;
-}
-
-std::string combine_values2(const nlohmann::json &json_obj) {
-    std::string combined;
-    for (const auto &item : json_obj.items()) {
-        combined += ToString2(item.value()) + " ";
-    }
-
-    if (!combined.empty()) {
-        combined.pop_back();
-    }
-    return combined;
-}
-
-inline std::vector<std::string> ConstructPrompts2(std::vector<nlohmann::json> &unique_rows, Connection &con,
-                                                  std::string prompt_name, int model_max_tokens = 4096) {
-    inja::Environment env;
-
-    auto query_result = con.Query(
-        "SELECT prompt FROM flockmtl_config.FLOCKMTL_PROMPT_INTERNAL_TABLE WHERE prompt_name = '" + prompt_name + "'");
-
-    if (query_result->RowCount() == 0) {
-        throw std::runtime_error("Prompt not found");
-    }
-
-    auto template_str = query_result->GetValue(0, 0).ToString();
-    auto row_tokens = Tiktoken::GetNumTokens(template_str);
-    auto max_length_values = GetMaxLengthValues2(unique_rows);
-    auto combined_values = combine_values2(max_length_values);
-    row_tokens += Tiktoken::GetNumTokens(combined_values);
-
-    std::vector<std::string> prompts;
-
-    if (row_tokens > model_max_tokens) {
-        throw std::runtime_error("The total number of tokens in the prompt exceeds the model's maximum token limit");
-    } else {
-        auto template_tokens = Tiktoken::GetNumTokens(llm_filter_prompt_template);
-        auto max_tokens_for_rows = model_max_tokens - template_tokens;
-        auto max_chunk_size = max_tokens_for_rows / row_tokens;
-        auto chunk_size = max_chunk_size > static_cast<int>(unique_rows.size()) ? static_cast<int>(unique_rows.size())
-                                                                                : max_chunk_size;
-        auto num_chunks = static_cast<int>(std::ceil(static_cast<double>(unique_rows.size()) / chunk_size));
-
-        for (int i = 0; i < num_chunks; ++i) {
-            nlohmann::json data;
-            data["prompts"] = template_str;
-
-            for (int j = 0; j < chunk_size; ++j) {
-                data["rows"].push_back(unique_rows[i + j]);
-            }
-
-            auto prompt = env.render(llm_filter_prompt_template, data);
-            prompts.push_back(prompt);
-        }
-    }
-
-    return prompts;
-}
-
-static void LlmFilterScalarFunction2(DataChunk &args, ExpressionState &state, Vector &result) {
+static void LlmFilterScalarFunction(DataChunk &args, ExpressionState &state, Vector &result) {
     Connection con(*state.GetContext().db);
     CoreScalarParsers::LlmFilterScalarParser(args);
 
@@ -121,7 +34,8 @@ static void LlmFilterScalarFunction2(DataChunk &args, ExpressionState &state, Ve
 
     auto tuples = CoreScalarParsers::Struct2Json(args.data[2], args.size());
 
-    auto prompts = ConstructPrompts2(tuples, con, args.data[0].GetValue(0).ToString(), model_max_tokens);
+    auto prompts = ConstructPrompts(tuples, con, args.data[0].GetValue(0).ToString(), llm_filter_prompt_template,
+                                    model_max_tokens);
 
     nlohmann::json settings;
     if (args.ColumnCount() == 4) {
@@ -149,7 +63,7 @@ static void LlmFilterScalarFunction2(DataChunk &args, ExpressionState &state, Ve
 }
 
 void CoreScalarFunctions::RegisterLlmFilterScalarFunction(DatabaseInstance &db) {
-    ExtensionUtil::RegisterFunction(db, ScalarFunction("llm_filter", {}, LogicalType::VARCHAR, LlmFilterScalarFunction2,
+    ExtensionUtil::RegisterFunction(db, ScalarFunction("llm_filter", {}, LogicalType::VARCHAR, LlmFilterScalarFunction,
                                                        nullptr, nullptr, nullptr, nullptr, LogicalType::ANY));
 }
